freifunk-common: split off "freifunk-common-olsr" package
Split the OLSRv1 specific things from the "freifunk-common" into a seperate package.
To keep the current package dependencies, the new package "freifunk-common-olsrv1"
must be included, which will include the basic "freifunk-common" package. For non-
olsrv1 setups only the "freifunk-common" package is needed, this way there is no
dependency to the olsr.

Signed-off-by: Sven Roederer <devel-sven@geroedel.de>