owm: drop runtime dependency to rpcd
the owm commandline tool and the common library depending on a running rpcd.
This dependency is caused by using the luci.model.uci package. As the tool runs
from the commandline anyway, switch to the uci package which provides the same
functionality with nearly the same API but without needing the rpcd.

For the commandline tool the switch from package "luci.model.uci" to "uci" can
be done straight forward. The library uses the get_first() call which is not
present in the replacement package. We switch to get() which should usually
return the same data as before, as there have been never a 2nd "olsrd" section
in the uci-file.
This also reduces the dependencies of the commandline tool to teh LuCI-framework.

Signed-off-by: Sven Roederer <freifunk@it-solutions.geroedel.de>